+++ src/dexterity/localroles/__init__.py
@@ -1,5 +1,6 @@
 # -*- extra stuff goes here -*-
 
+from plone import api
 from zope.i18nmessageid import MessageFactory
 
 import logging
@@ -10,6 +11,7 @@ PMF = MessageFactory('plone')
 
 
 logger = logging.getLogger('dexterity.localroles')
+HAS_PLONE_6 = int(api.env.plone_version()[0]) >= 6
 
 
 def initialize(context):