+++ CHANGES.rst
@@ -5,8 +5,9 @@ Changelog
 --------------------
 
 - Added Plone 6.1 version in buildout.
-  [chris-adam]
-
+  [chris-adam, sgeulette]
+- Added `dexterity.localroles.related_parent_with_portal` utility to include Plone Site in related parent search.
+  [sgeulette]
 
 2.0.0a (2023-11-28)
 -------------------